stagifyapi/requirements.txt
django==3.1.3
psycopg2-binary==2.8.6
django-cors-headers==3.5.0
django-storages==1.10.1
django-storages[azure]==1.10.1
